[skip ci] wip
pausing here - at this point it's actually non-trivial to have the
points on the x axis be non-linear; this is due to the fact that we
scale the chart to the available pixels for performance reasons.

I'm not sure how to easily map back from a pixel position back to the
correct index without resorting to a lot of hacks...

Saving now - so if I ever look at this later I might have some ideas.

Was for #3.